chatbot-api/pyproject.toml
[tool.poetry]
name = "chatbot-api"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
httpx = "^0.27.2"
faiss-cpu = "^1.8.0.post1"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
transformers = "^4.45.1"
torch = "^2.4.1"
python-multipart = "^0.0.10"
numpy = "^1.26.4"
sentence-transformers = "^3.1.1"
psutil = "^6.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"

[tool.pytest]
asyncio_mode = "auto"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
isort = "^5.13.2"
mypy = "^1.11.2"
pytest = "^8.3.3"
anyio = "^4.6.0"
pytest-asyncio = "^0.24.0"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.3"
twisted = "^24.7.0"
sqlalchemy-utils = "^0.41.2"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"